# Progression on rotationed averaging
New logic has been added so that results are not output to STDOUT
until the last orientation has been run.

WORK IN PROGRESS: Results are being averaged, but further testing is
required. Logic needs to be added to bypass component orientation output
as this is meaningless for these types of simulations. Also, the command
line switch may be removed and replaced with input in the Choice file.