Adds 2015 NBC analysis folder and scripts
This application folder contains Perl scripts for interfacing with HTAP to run
2015 NBC ruleset simulations with htap-prm.rb. The script relies on a location
cross-referencing XML database which maps climate locations to HVAC and DHW
fuel types, and locations to archtype subgroups. The current HVAC and DHW cross-
referencing maps common system fuel types for each location as determined by
OEE's CEUD.

The script calls htap-prm.rb, collects energy output, and stores the results in
an XML database for future processing.

NOTE: The script points to XML database ArchInfo.xml which is not included in this
commit. This will be committed following official release of the EGH-Archs database.