Added exhaust side only ventilation for NBC ruleset
NOT FOR PRODUCTION: The HRVonly option was updated to include a procedure to create a
whole-dwelling exhaust-side ventilation system. Previous simulations without an HRV assumed
a balanced system. This has been shown to consume more energy than exhaust side (up to 12% in
climate zone 8). The HRVonly option may be used to create an exhaust side ventilation system by
setting the IsActive attribute to 'utility'. When this is invoked, the follwing attributes are
ignored: Rating1, Rating2, Rating3, FanPower2. This is ineligant, however, is currently fit for
purpose. The NBC perl script has also been updated to reflect these changes.