Changes to NBC ruleset for HVAC and DHW setting
This ruleset scans the archetypes HVAC and DHW systems and applies a code-minimum
system using the same fuel as the base model. There have been some issues encountered
for archetypes using solar and propane as their base fuels. For space heating there
was some issues with setting an air heat pump as the primary space heating system. This
has been modified so that electric baseboard heaters are used instead as a temporary
measure. For domestic hot water if the system isn't oil, electric, or natural gas (like
solar or propane) the same fuel that the heating system uses is assumed for the hot water.