Added new value to Opt-GenericWall_1Layer_definitions
The new value `MappedIntFloorHeader` has been added to `Opt-GenericWall_1Layer_definitions`. This allows the user
to specify a `Opt-FloorHeaderIntIns` to be used with a `Opt-GenericWall_1Layer_definitions` option. The current
implemented precedence is ruleset < choice < MappedIntFloorHeader option. A warning is issued to the substitute log
if MappedIntFloorHeader overrides either a choice or rulest option.