Git version stamping (#81)
* Added definition that returns git branch and revision number

This initial commit adds a definition which returns the branch and short commit SHA1. It has been
tested in PowerShell using the commented-out call and print statements in htap-prm.rb. The intention
is to have this information included in the HTAP-prm-output.json files.

* Patch to track h2k file versions alongside simulation results.

* Changes to include version info in results output

* Changes to include version info in results output

* This patch ensures that git informaton is always returned for
the script location (e.g. C:\HTAP\), which may differ from the
invoked locaiton.