Added Skylights to NBC Ruleset. Added new option Opt-FoundationSlabBelowGradeAboveFrost. Opt-AboveGradeWalls now applies to ponywalls
Skylight U-values have been added to the NBC ruleset.

The new option Opt-FoundationSlabBelowGradeAboveFrost has been introduced. This slab-below-grade insulation is applied to
walkouts and any basement whose depth is above the local frostline depth. Frost depths per location are specified in constants.rb in the
global hash $Frostline. Those with unknown depth are specifed to 999 and it will be assumed that all basements are below the frostline. Tested
with the following archetypes in Climate Zone 6 (OTTAWA):
- ERS-7032: Basement below frost line
- ERS-7972: Basement above frost line
- ERS-1727: Walkout foundation with pony wall
- ERS-1552: Basement foundation with pony wall